Add: eviction of least recently used upon maxSize threshold reached

+++ kvs/options.go
@@ -1,7 +1,7 @@
 package kvs
 
 type StoreOptions struct {
-	MaxSize uint64
+	MaxSize int
 }
 
 type OptionSetter func(o *StoreOptions)
@@ -16,7 +16,7 @@ func LoadOptions(opts ...OptionSetter) *StoreOptions {
 	return newOptionsInstance
 }
 
-func WithMaxRAMSize(size uint64) OptionSetter {
+func WithMaxRAMSize(size int) OptionSetter {
 	return func(o *StoreOptions) {
 		o.MaxSize = size
 	}
